sensor-binder/requirements.txt
pymongo
kafka-python
dnspython
